Move helpers and constants to TokenLoansContent from LoansTokenActiveTable
