+++ src/pages/tokenLending/LoansTokenPage/LoansTokenActiveTable/TableCells/RefinanceTokenModal/OrderBook/cells.tsx
@@ -14,6 +14,7 @@ import {
 } from '@banx/components/TableComponents'
 
 import { core } from '@banx/api/tokens'
+import { calculateTokensToGet } from '@banx/pages/tokenLending/LoansTokenPage/TokenLoansContent/helpers'
 import {
   adjustTokenAmountWithUpfrontFee,
   caclulateBorrowTokenLoanValue,
@@ -24,8 +25,6 @@ import {
   getTokenDecimals,
   getTokenUnit,
 } from '@banx/utils'
-
-import { calculateTokensToGet } from '../../../helpers'
 
 import styles from './OrderBook.module.less'
